add Home component
